requirements.txt
openai>=1.12.0
faiss-cpu>=1.7.4
numpy>=1.26.0
sentence-transformers>=3.0.0
PyYAML>=6.0

